refactor: merge features
