--- src/app/[locale]/_old_client/_components/Code.tsx
@@ -1,13 +1,0 @@
-"use client";
-import { CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-
-export default function Code() {
-  return (
-    <div>
-      <CardHeader className="px-0">
-        <CardTitle>Code:</CardTitle>
-      </CardHeader>
-      <CardContent className="px-0 py-6"></CardContent>
-    </div>
-  );
-}
