+++ src/app/[locale]/client/_components/BodyEditor.tsx
@@ -1,17 +1,74 @@
-import { FC } from "react";
+import { FC, useState } from "react";
+import { Select } from "@/components";
+import { Button } from "@/components/ui/button";
 
-const BodyEditor: FC<{ decodedBody?: string }> = ({ decodedBody }) => {
-  const onChange = () => {
-    // can do it controlled or anymore...
+const selectOptions = ["JSON", "text"];
+
+const Body: FC<{ body?: string; onBodyChange: (body: string) => void }> = ({
+  body,
+  onBodyChange,
+}) => {
+  const [text, setText] = useState<string>(() => body || "");
+  const [error, setError] = useState<string | null>(null);
+  const [mode, setMode] = useState("JSON");
+
+  const handleTextChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+    const value = e.target.value;
+    setText(value);
+
+    if (mode === "JSON") {
+      try {
+        JSON.parse(value);
+        setError(null);
+      } catch (err) {
+        setError(err instanceof Error ? err.message : "Invalid JSON");
+      }
+    } else {
+      setError(null);
+    }
   };
 
-  // eslint-disable-next-line
-  const onSubmit = () => {
-    //todo: change path
-    // similar to url input
+  const handleChangeMode = (event: React.ChangeEvent<HTMLSelectElement>) => {
+    const value = event.target.value;
+    setMode(value);
+
+    if (value === "text") {
+      setError(null);
+    }
   };
-  // use tool json-editor or similar instead textarea
-  return <textarea value={decodedBody} onChange={onChange}></textarea>;
+
+  return (
+    <div className="container mx-auto">
+      <div className="mb-4 flex items-end justify-between">
+        <Select
+          label="Mode"
+          options={selectOptions.map((value) => ({ label: value, value }))}
+          value={mode}
+          onChange={handleChangeMode}
+        />
+        <Button
+          onClick={() => {
+            onBodyChange(text);
+          }}
+        >
+          Ok
+        </Button>
+      </div>
+
+      <textarea
+        value={text}
+        onChange={handleTextChange}
+        className="w-full h-60 p-2 border rounded font-mono text-sm"
+        spellCheck="false"
+      />
+
+      {error && (
+        <div className="mt-2 p-2 bg-red-100 text-red-700 rounded">
+          Error: {error}
+        </div>
+      )}
+    </div>
+  );
 };
 
-export default BodyEditor;
+export default Body;
