+++ src/app/[locale]/client/_components/Client.tsx
@@ -1,0 +1,129 @@
+"use client";
+
+import { Link, usePathname, useRouter } from "@/i18n/navigation";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import Headers from "./HeadersHandler";
+import InputUrl from "./InputUrl";
+import Code from "./CodeView";
+import Body from "./BodyEditor";
+import { useTranslations } from "next-intl";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import ResponseView from "./ResponseView";
+import { Button } from "@/components/ui/button";
+import { useParams, useSearchParams } from "next/navigation";
+import { useState } from "react";
+
+export default function ClientPage() {
+  const t = useTranslations("Client_Page");
+  const pathname = usePathname();
+  const router = useRouter();
+  const params = useParams<{ request: string[] }>();
+  const pathnameArr = pathname.split("/");
+  const searchParams = useSearchParams();
+  const query = Object.fromEntries(searchParams.entries());
+
+  const [method, url, body] = params.request;
+
+  const onMethodChange = (newMethod: string) => {
+    router.replace(
+      { pathname: pathname.replace(method, newMethod), query },
+      { scroll: false },
+    );
+  };
+  const urlDecoded = url && atob(decodeURIComponent(url));
+  const onUrlChange = (url: string) => {
+    pathnameArr[3] = btoa(url);
+    router.replace(
+      {
+        pathname: pathnameArr.join("/"),
+        query,
+      },
+      { scroll: false },
+    );
+  };
+
+  const bodyDecoded = body && atob(decodeURIComponent(body));
+
+  const onBodyChange = (text: string) => {
+    pathnameArr[4] = btoa(text);
+    router.replace(
+      {
+        pathname: pathnameArr.join("/"),
+        query,
+      },
+      { scroll: false },
+    );
+  };
+
+  const onQueryChange = (query: string) => {
+    router.replace(`?${query}`, { scroll: false });
+  };
+
+  //eslint-disable-next-line
+  const [response, setResponse] = useState<Promise<Response> | undefined>();
+
+  const onGo = () => {
+    //
+    //
+    // TODO: variables insertion
+    // TODO: create route handlers and
+    // const response = fetch(`/api/request?url=${url}`, {
+    //   method,
+    //   headers: _headers,
+    //   body,
+    // });
+    // setResponse(response)
+  };
+
+  return (
+    <main className="flex flex-col gap-8 py-4 items-center">
+      <section>
+        <Card className="mx-6 w-4xl">
+          <CardHeader>
+            <CardTitle className="flex flex-col items-center">
+              {t("REST_Client")}
+            </CardTitle>
+          </CardHeader>
+          <CardContent className="flex flex-col gap-8">
+            <div className="flex items-end justify-between">
+              <InputUrl
+                {...{ method, onMethodChange, onUrlChange, url: urlDecoded }}
+              ></InputUrl>
+              <Button onClick={onGo}>Go!</Button>
+            </div>
+            <Tabs defaultValue={"headers"}>
+              <TabsList>
+                <TabsTrigger value="headers">Headers</TabsTrigger>
+                <TabsTrigger value="body">Body</TabsTrigger>
+                <TabsTrigger value="code">Code</TabsTrigger>
+              </TabsList>
+              <TabsContent value="headers">
+                <Headers {...{ query, onQueryChange }} />
+              </TabsContent>
+              <TabsContent value="body">
+                <Body body={bodyDecoded} onBodyChange={onBodyChange} />
+              </TabsContent>
+              <TabsContent value="code">
+                <Code />
+              </TabsContent>
+            </Tabs>
+          </CardContent>
+        </Card>
+      </section>
+      <section>
+        <Card className="mx-6 w-4xl">
+          <CardHeader>
+            <CardTitle className="flex flex-col items-center">
+              <ResponseView response={response} />
+            </CardTitle>
+          </CardHeader>
+          <CardContent></CardContent>
+        </Card>
+      </section>
+
+      <Link href="/variables" className="mt-8 p-4">
+        <button>Go to Variables</button>
+      </Link>
+    </main>
+  );
+}
