+++ src/app/[locale]/client/_components/CodeView.tsx
@@ -1,7 +1,13 @@
-import { FC } from "react";
+"use client";
+import { CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 
-const CodeViewer: FC<unknown> = () => {
-  return <div></div>;
-};
-
-export default CodeViewer;
+export default function Code() {
+  return (
+    <div>
+      <CardHeader className="px-0">
+        <CardTitle>Code:</CardTitle>
+      </CardHeader>
+      <CardContent className="px-0 py-6"></CardContent>
+    </div>
+  );
+}
