+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./Modal";
+export * from "./Table";
+export * from "./Select";
